requirements.txt
flask==2.3.3
graphene==2.1.9
flask-graphql==2.0.1
requests==2.31.0
streamlit==1.28.1
python-dotenv==1.0.0

